Arithmetic: options to add metadata (name, units, comment) in output
Until now, after a user finished their desired operations on an image and
generated and output with Arithmetic, if they wanted to add metadata to the
image, they would have to use a separate call to the Fits program and add
them. This extra command was annoying and could even discourage adding
metadata.

With this commit, the Arithmetic program has three new options to add a
name ('--metaname'), unit ('--metaunit') and a comment ('--metacomment') to
the output HDU.