Fits: units of a HDU are also printed when no option given
Until now, when no option was given, the table printed by the Fits program
(that contains metadata about the HDUs) didn't show the units metadata! But
the FITS standard does specify a special keyword for units ('BUNIT'), and
it is commonly available from some datasets.

With this commit, a new units column has been added for this scenario. HDUs
that contain arrays (like 2D images or 3D cubes; not tables), and have the
'BUNIT' keyword, will have a non-blank value in this column.