MakeCatalog: --skystd is new name for the old --std
Until now, we were using the '--std' name for the sky standard deviation
(using the sky standard deviation image). However, given that we have
columns like '--mean', '--median' or etc (which report the respective
statistic of the values image), the name '--std' could lead to the user
thinking this is the same (using the values image pixels).

With this commit, the old '--std' option has been renamed to '--skystd' to
avoid such confusion (which is indeed a bug).

This bug was reported Raul Infante-Sainz.

This fixes bug #62685.