radial-profile: setting spatial resolution to zero
Until now, the spatial resolution used in the radial profile script was not
specified.  Consequently, the default value was used in the call to
MakeCatalog for measuring the radial profile
('--spatialresolution=2'). This parameter is related with the uncertainty
in the measurement of the area. But since for the radial profile script we
are considering fixed apertures, the errors of this apertures are zero.

I found this issue when obtaining the surface brightness error
('--sberror') of simulated profiles. For each radius of a bunch of
simulated sersic profiles, I compared the scatter (STD) of the surface
brightness measurements ('--surfacebrightness') with respect to the
measured surface brightness error ('--sberror'). With this experiment, I
realized that the measurements are compatible with the expected results
when the '--spatialresolution' is 0.0.

With this commit, the parameter '--spatialresolution' has been set to zero
for the radial profile script. We are still investigating about how this
parameter has to be set in other/general situations.

This bug was fixed with the help of Mohammad Akhlaghi.

This fixes bug #62694.