MakeCatalog: new --std column for STD of pixels in values image
Until now, there no way to estimate the standard deviation of values of
each labeled region (objects or clumps) using MakeCatalog.

With this option, the newly added '--std' option will do this job. Until a
few commits ago, this was the name for the sky standard deviation, so the
NEWS file was also updated in the "changes" section.

In the process, the calculation of the standard deviation from the sum,
sum-of-squares and number of elements has been taken into a new Gnuastro
library function in 'statistics.h': 'gal_statistics_std_from_sums'