Library (type.h): integer strings force-parsed as base 10
Until now, when parsing integer strings as integers with 'strtol', the base
was set to '0' (which means that integers starting with '0' should be
interpreted as base 8 and those starting with '0x' should be parsed as
base-16). While bases, are very common in programming and computer science,
they are rarely used in data analysis. In fact, some pipelines do '0'
padding to make sure that the number has a certain number of digits (for
example '0015').

We had already accounted for this when parsing text tables (in 'lib/txt.c')
by setting the base of 'strtol' to 10. But in 'gal_type_string_to_number'
the base was '0'. As a result, when reading FITS keyword values with
integers starting with '0', the program would either crash (because '8' or
'9' are not defined in base-8), or return unreasonable values since the
base was different.

With this commit, 'strtol' now uses base 10 and this fixed the problem.

I also noticed that this function was causing a crash when the integer
number couldn't be read from the string. In such cases, it should return
NULL. So that has also been corrected.

Furthermore, the documenation of 'gal_type_string_to_number' in the book
now fully describes the problem of implicit type conversion for binary
operators and the shorter description in the comments above
'arithmetic_binary_int_sanity_check' have been removed and point the reader
to that part of the book.

This bug was found by Manuel Sánchez-Benavente.

This fixes bug #62710.