Book: the quick start section now focuses on Lzip
Until now, in the quick start section we were talking extensively about
both the Gzip and Lzip formats. But this could be confusing for some
readers. Although the Gzip'd tarball is available for official releases,
its not available for the alpha releases and generally due to its larger
volume and weaker archival standards, it is not recommended.

With this commit, the description focuses on Lzip, but just briefly notes
that Gzip is available for official releases.

This issue was raised by Peter Teuben.