Library: Added source files for Python wrappers
Added the python.h and python.c source files in the lib folder
for providing functionality required in building the Python
extension modules fro Gnuastro.

Also, added the Python extensions modules for cosmology library
and the setup utility for the same. This will allow the use of
cosmology library functions through Python3.

Also, added a --with-python option in the configure file to enable
users to opt for a build with/without the Python extensions.

Fixed a few typos in the doc as well.