Library (fits.h): Table info is read when END key not read by CFITSIO
Until now, the 'gal_fits_tab_info' function would only stop parsing the
headers of the input FITS file when CFITSIO returned the 'END'
keyword. However, as reported by Hilderic Browne in [1], for some FITS
files, strangely CFITSIO doesn't return the 'END' keyword. As a result,
Table would fall into an infinite loop and would never stop!

After finding that the problem originates from CFITSIO, it was immediately
reported to the CFITSIO maintainers at 'ccfits@heasarc.gsfc.nasa.gov' and
they are looking into it.

With this commit, a work-around has been implemented (which is necessary
because even after the fix in CFITSIO, it may take many years for many
users to update it): immediately after attempting to read the next keyword
in a FITS file, a check has been added for the 'status' output of
CFITSIO. If we have indeed passed 'END', the 'status' will be
'KEY_OUT_BOUNDS'. In this case, we will manually write 'END' and let the
loop stop as before.

While checking that same file, I noticed that while reading blank columns,
we aren't setting the string blank columns properly (instead of directly
passing the pointer to the blank string, we were passing the pointer to a
pointer!). With this commit, this bug has also been fixed.

[1] https://lists.gnu.org/archive/html/bug-gnuastro/2022-07/msg00005.html