Library (txt.c): better error message when _:_:_ can't be parsed
Until now, the error message that was printed when a celestial coordinate
in the format of '_:_:_' was input wasn't clear: it didn't clarify that it
is important to have metadata to have a string column.

With this commit, the last part of the error message has been extended to
make it clear that metadata is necessary for the 'ra-to-degree' or
'dec-to-degree' operators (and that it is easier to simply use the
unambiguous '_h_m_' and '_d_m_' formats).

This was suggested by Manuel Sánchez-Benavente.