Library (txt.h): last columns with string format correctly read
Until now, when the last column of a table was a string, and the specified
width (in the metadata) is larger than the width of the string (before the
new-line character), Table would print unreasonable characters after the
main column value. For example see this command

    printf "# Column 1: RA [s, str20]\n%s" 03:24:15 | asttable

In the command above, the string metadata has given a width of 20
characters, while the actual string is 8 characters! So some extra
non-sense characters are printed after the output! If we set the metadata
to 'str8' or 'str9' there is no problem (the 'str9' will also copy the
string termination character).

With this commit, the cause of the problem has been found and fixed. The
main problem was that we were only copying the shorter content, but we
weren't setting the string termination character in the end of the copied
number of bytes.

I also noticed that for allocating new space, we were directly using the
low-level 'malloc' instead of the high-level 'gal_pointer_allocate' which
is also cleaner. We now use the high-level allocation function.

This bug was found based on a discussion with Manuel Sánchez-Benavente.

This fixes bug #62720.