radial-profile: larger first azimuth angle will use outer part
Until now, we were just able to generate an azimuthal profile with an
initial angle smaller than the final angle. If we needed to have a
symmetric wedge along the semi-major axis, in which the initial angle is
bigger than the final angle for exmaple from 350 to 10, the radial profile
script would crash.

With this commit, this bug is fixed by adding a conditional statement in
the call to 'astarithmetic'. In effect, this has also increased the
functionality of the radial profile script: because the user can now ignore
parts of the radial profile and use the reset of the azimuthal angles.

In the process, the following issues have also been corrected:

  - When the azimuthal angles are used, the central pixel is effectively
    ignored because it is no longer a "hole" and gets connected to the
    regions outside the radial range. So the step to fill the central hole
    and increment the radial distances (that will be corrected in the final
    profile) is now done before the azimuthal correction.

 - The azimuthal profile (to select pixels in the desired azimuthal range)
   is now generated in float32 numerical type, not int16! This allows more
   accurate selection of the azimuthal range, especially in larger
   profiles.

This fixes bug #62721

The documentation and the parts relating to the central pixel have been
added by Mohammad Akhlaghi.