Library (fits.h): table metadata parsed by returned status of CFITSIO
Until now, we would check for the 'END' keyword to stop parsing all the
FITS keywords (when it was necessary to extract table metadata). However,
due to some intricacies in CFITSIO [1], this could cause problems in
situations when the keyword before 'END' is empty (see the attached file in
[2]): the 'END' keyword was never returned by 'fits_read_keyn', so the loop
in 'gal_fits_tab_info' would never terminate.

Upon finding the problem, a dirty work-around was implemented in Commit
7d117daa2d0 ("Library (fits.h): Table info is read when END key not read by
CFITSIO"). But that was not an elegant solution because it would replace a
fake 'END'.

With this commit, following the advice from Craig in [2], we are no longer
checking for the 'END' keyword at all. Instead, we check the return status
of 'fits_read_keyn' and abort the loop when that status is non-zero. If the
non-zero status is 'KEY_OUT_BOUNDS' then everything is fine (we have
finished parsing all the keywords). Otherwise, 'gal_fits_tab_info' aborts
the Gnuastro program calling it, and reports the unexpected CFITSIO return
status error message.

[1] Reply of Craig Gordon (from CFITSIO) on this problem:

    The reason the fits_read_keyn() call for key number 284 returns an
    empty string is because there really is just a blank line at that
    position.  The END keyword is actually located 1 line after that, at
    what would be keyword position 285.

    It is not the usual case to have blank lines in between the last real
    keyword in the header and the END keyword, which is why you haven't run
    into this with other files.  But blank lines are allowed here by the
    FITS standard, so this is still a valid file.  (If new keywords are
    added to the header, CFITSIO may insert them into this blank space.)

    The problem you're running into is that while END exists at key position
    285, CFITSIO functions such as fits_read_keyn() and fits_find_nextkey()
    only let you access one position beyond the last real keyword in the header
    (which you can find by functions such as fits_get_hdrspace() and
    fits_get_hdrpos()).  So if you tried replacing 284 with 285 in your
    example, fits_read_keyn would return an error.

    This brings up the question of whether it is necessary for Gnuastro to
    actually access the END keyword. In any case it's probably not a good
    idea to use the location of the END keyword as the criteria for
    stopping a keyword-reading loop. You could just cut it off at the
    position of the last non-END keyword, as given by fits_get_hdrpos and
    fits_get_hdrspace.  Or you could also loop through all the keys by
    continually calling the fits_find_nextkey function until it returns a
    bad status.

[2] https://lists.gnu.org/archive/html/bug-gnuastro/2022-07/msg00005.html