MakeCatalog: warning will be printed for small values to --upnum
Until now, when the value given to '--upnum' was smaller than 20,
MakeCatalog would abort since the result of all upper-limit measurements
will not be reliable. However, in some contexts, the upper-limit
measurements aren't relevant and the user wants the random positions for
other things (like using '--checkuplim' to get positions where there is no
signal for completeness tests).

With this commit, instead of aborting, MakeCatalog will print an
informative warning to let the user know that the upper-limit measurements
won't be reliable, and it will continue with the rest of the processing. As
a result, when we want a small number of random positions, MakeCatalog can
be used.

This was suggested by S. Zahra Hosseini Shahisavandi and Ignacio Trujillo.