Book: better colorbar in 2D histogram example with PGFPlots
Until now, the colorbar used in the demo histogram of the Statistics
section of the book was based on the shades of blue (which is not too clear
for the eye (the contrast between the values was very low).

With this commit, a different colorbar is defined that is much more clear,
with many colors and much more contast. I also noticed that even though the
commands on how to build the LaTeX source were given in comments within the
demo LaTeX source, there was no description in the main text of the book!
Therefore, some explanation on the necessary commands has also been added.

Also, a spell-check has been done in some parts of the book.