Library (units.h): using signbit() instead of <0.0f
Until now, while parsing the sexagesimal RA or Dec strings, to check for
negative values, we were simply using '<0.0f'. However, this test will fail
when given a negative zero (or '-0.0', which can happen in sexagesimal
coordiantes in small declinations for example '-00d12m34'). As a result, a
command like below would effectively ignore the sign in its output:

    $ echo -00h12m34 | asttable
    3.141666667

With this commit, the two 'gal_units_ra_to_degree' and
'gal_units_dec_to_degree' functions now use the 'signbit()' function of the
C library to check for the "sign" bit (and thus identify if it is positive
or negative). This allows them to properly detect a negative zero and
account for it like all negative numbers.

This bug was reported by Manuel Sánchez-Benavente.

This fixes bug #62784.