psf-stmap: nan stamp if star is completely outside of field of view
Until now, if a star was completely outside of input image and did not have
any overlap with the input image, the script would abort and print
error. However, later, if a normalization value was found to be NaN, it
would produce a fully NaN-valued image. This inconsistant exit behavior
makes it hard to use the output in pipelines.

With this commit, a unified output is always created upon aborting: an
output image is going to be produced, but an image that is fully
NaN-valued. An informative warning is also printed just before the program
ends in both cases.

While inspecting Sepideh's commit, Mohammad noticed the following two
issues that are also implemented in this commit:

 - The Fits program can't write NaN-valued FITS keywords (which was why we
   were not including the 'NORMVAL' keyword when the normalization value
   was NaN until now). Therefore, the proper check was added within the
   FITS program so when a given value is NaN, it prints an empty-valued
   keyword (permitted by the FITS standard, and CFITSIO).

 - The Library's 'gal_blank_is' function wasn't prepared for checking
   strings that aren't "n/a" (the blank string value)! If the string was
   something else, it would go to the next 'case' element which was for
   complex numbers and produce confusing error messages.

This fixes bug #62794.