NEWS: preparations for Gnuastro 0.18
Until now, we were developing Gnuastro 0.17, but Gnuastro 0.18 is finally
ready for release.

With this commit, the necessary parts of Gnuastro have been updated for
Gnuastro 0.18's release: most importantly the NEWS file.

In the process, with the help of to Irene Pintos Castro and Zahra Sharbaf a
test build was also done on two macOSs. Irene's macOS printed a warning
that has been corrected with this commit.