NEWS: added titles for new changes
Until now, the top entry on the NEWS file was for version 0.18. But that
version is already public and all new development needs to be documented in
a new title entry.

With this commit, a new title has been added for all changes until next
release. Also, the SONAME of the dynamic library has been incremented in
preparations that will be made there (until the next version). Finally, the
notice at the start of the book has been added again (informing readers
that this is not the official release).