ConvertType: now possible to draw vector graphics marks on EPS/PDF
Until now, there was no easy interface to mark certain objects or add text
under the marks in the publication-ready outputs (PDF or EPS). The only way
to have high quality (vector graphics) marks and text over the image in
your report was to use PGFPlots to draw them within the actual
paper. However, within LaTeX the units are in centimeters or page width,
not pixels and converting RA/DEC to the LaTeX coordinates for each image
was very annoying.

With this commit, it is now possible to give a "Marks" catalog to
ConvertType. Using that catalog, ConvertType will draw shapes from the
given coordinates in the table (and optionally add text under the
mark). The user also has freedom to select 140 colors from the extended Web
color formats, set the line width, rotation and many other ways to
customize the marks.

In order to do this, many options have been added to ConvertType and to
easily adopt/learn these new features, two new sections have been added to
the general tutorial.

In the process, several other things have been done with this commit:

 - Based on feedback from many users, the general tutorial is now the first
   section of the tutorials chapter. The "Sufi simulates a detection"
   tutorial has been moved to the fourth section (because it is about
   simulating data, while the first three are about using existing data).

 - Mathematical (like 'pi') and Physical constants (like the speed of
   light) are now available in the Arithmetic library and usable in Table's
   column arithmetic and also the Arithmetic program.

 - Four new fuctions have been added to the Arithmetic library to convert
   surface brightness to and from magnitudes and counts.

 - The Fits program now has a new option called '--pixelareaarcsec2' which
   will return the pixel area in units of arc-seconds squared.

 - Many new library functions and macros have been added to enable easy
   addition of marks.