Python (setup, cosmology): Added the setup script and cosmology module
Until now, we added the necessary checks and utilities needed for building
the Python package. Here we add the setup script for building the Gnuastro
Python package with the cosmology module and a new section in the book.

With this commit, users will now be able to import the Python package
and use the cosmology module. The package is built by applying Python
wrappers around existing Gnuastro library functions. Currently, only
the cosmology module has been added. The setup script also supports
VPATH builds, i.e separate build and source trees.

Added instructions in the Makefile to only build the Python package only
if the checks for Python 3 and NumPy succeed.

Also added a new section in the book describing the Python package, it's
need, installations instructions, some examples and description of the
extension modules.