Fits: --printkeynames accounts for empty lines before END
Until now, when the '--printkeynames' option of the Fits program was run on
a FITS file that had empty keywords before the 'END' keyword, it would
never abort (falling into an infinite loop). For the causes of this issue,
see the description of Commit 5f1d2c735e75d, titled "Library (fits.h):
table metadata parsed by returned status of CFITSIO".

With this commit, a similar fix to that commit has been implemented:
instead of stopping the loop when we hit the 'END' keyword, we stop when
CFITSIO returns a non-zero 'status'.

This bug was reported by Pedram Ashofteh Ardakani.

This fixes bug #62861.