Fits: added sanity check when reading keyword names
Until now (the previous commit), we would just stop the loop (that parses
all the keywords for their names) when the 'status' was non-zero. In
expected scenarios (where the only non-zero exit status is
'KEY_OUT_BOUNDS', this is fine. But we hadn't accounted for any other
possible exit status (due to some unexpected error).

With this commit, a check has been added that will stop the loop if the
exit status is not zero or 'KEY_OUT_BOUND'.