Book: more links and clarifications in New to GNU/Linux
Until now, the "New to GNU/Linux?" section wasn't too complete (not
pointing to some relevant GNU links). During an email discussion with
Richard Stallman, he pointed to the following links that would be useful in
this section:

Richard also had the following comment about the car carburetor example
(which was taken from [3]): "I think that's a little too strong as a
put-down of Linx. The kernel is a major system component, much bigger in an
OS than a carburetor is in a car.  Many cars don't even have carburetors,
but every system needs something comparable to the kernel."

With this commit, the links of [1] and [2] have been added and we now
compare the kernel to the engine of a car, with some further explanation
that the engine along won't get you from A to B: you still need wheels, a
chassis, seats, a wind shield and etc.

[1] https://gnu.org/philosophy
[2] https://www.gnu.org/gnu/the-gnu-project.html
[3] https://www.gnu.org/gnu/gnu-users-never-heard-of-gnu.html

The added part on the Windows Subsystem for Linux (WSL) was confirmed with
the help of Marjan Akbari.