Installed scripts: LC_NUMERIC=C set to prevent commas in floats
Until now, in the installed scripts, 'export LANG=C' was set to prevent
commas instead of floating points. However, we realized that even with this
option, there were some situations where commas appeared. This is because
there are several different local environment variables and depending on
the particular configuration of the OS they can be different.

For example 'LANG' is the language to print human-text, but 'LC_NUMERIC' is
the format to print numbers, and they can be different! The relevant one
for this purpose was is 'LC_NUMERIC'. By setting it to 'C' (for
programmers), floating points won't be displayed as commas, but programs
that print human text in the user's default language, will continue to do
so with on inconvenience for the user.

With this commit, 'export LC_NUMERIC=C' has been replaced with 'export
LANG=C' at the beginning of all scripts. This fixes situations like
above. Also, to further make sure that the tests are done in a "clean"
environment, 'LC_NUMERIC' is also passed as an environment variable to all
the check scripts.

Finally, to make it easier for people to learn about this problem, the text
that was previously in a box in the book has been moved to a whole new
section in the "Common program behavior" chapter.

This issue was found by Teet Kuutma, due to a failure in the 'make check'
phase.