psf-stamp: sub-pixel centering is now applied
Until now, the 'astscript-psf-stamp' script would only crop the desired
region from the input image and leave the pixel grid untouched. This was
not good for calcluating the central parts of the PSF because the center of
the star could be in the edge of a pixel.

With this commit, when the center of stars are not in the center of the
pixel, at first the displacement of value will be caluculated and put then
warp the image based on the displacement value.

On the other hand, when the stars in the border of lefthand bottom this
displacement will be calculated mistakenly. But we add a condition when
object is in the lefthand bottom border caculate displacement correctly and
warp it.