psf-stamp: new --snthresh option to mask low S/N pixels
Until now, the 'astscript-psf-stamp' script would only optionally mask the
other clumps/objects in its output. However, another type of pixels that
users don't want are low signal-to-noise ratio pixels (which can degrade
the final stack's S/N when stars of different magnitudes are used.

With this commit, a new option has been added to this script for this
purpose: if the user gives this option a value, all pixels below that S/N
will be masked (and not harm the regions from brighter stars).

Also, until now, to specify the size of the stamps in these two PSF scripts
we would use the '--stampwidth' option. However, this new name is hard to
remember! Based on other gnuastro programs like Crop, users expect to have
the option name start with '--width*'. With this commit, this option has
been renamed to '--widthinpix' to be similar to a similarly named option in
Crop, while also being clear that the expected units are in pixels.

This was implemented after a discussion with Sepideh Eskandarlou and Raul
Infante-Sainz.