fits-view script: single ds9 colorbar by default
Until now, when DS9 was called with this script, it would load a separate
color bar for each separate image. But this would waste a lot of space when
viewing many images.

With this commit, by default a single color-bar is shown and the user is
able to have the old behavior (multiple color bars) with the
'--ds9colorbarmulti' option.

In the process, two other minor corrections were made to this script:

 - Some of the lines in this script were indented with a TAB! This has been
   corrected.

 - The list of options to DS9 was getting long and hard to read. So there
   DS9 options are now called on a one-line-per-option format.