Book: separate calls for mandatory and optional dependencies
Until now, there was a single package manager command (in all operating
systems) for all the optional, mandatory and non-Gnuastro (topcat and DS9)
dependencies. Especially after the addition of DS9 and Topcat, by using
those commands, MANY MANY extra dependencies of dependencies will be
installed with that single command (since DS9 and Topcat are GUI). This
could lead to the idea that Gnuastro also has many dependencies!

With this commit, we now have a separate calls to the operating system's
package manager for each class of dependencies. This will hopefully let the
users get a better feeling of what they need to install based on their
purpose, before installing Gnuastro.