Book: corrected description of demo commands at start of ConvertType
Until now, the ConvertType demo command for inserting multiple files for
each color channel at the start of the "Invoking ConvertType" section, was
saying "three plain text" files. However, in the example, the third was
actually a FITS file!

With this commit, the description is edited to say that the third is
actually a FITS file, showing that it is possible to merge different
formats for each color channel.

This issue was reported by Faezeh Bijarchian.