psf-scale-factor: --quiet option passed to internal astarithmetic
Until now, one of the 'astarithmetic' calls within this script wasn't being
passed the '--quiet' option. So when the user wanted to use the scale
factor value as a variabe, could not do it. Because '--quiet' option was
not active and a lot of extra things were printed, making it hard to
extract the calculated scale factor.

With this commit, the '$quiet' option is passed to that 'astarithmetic'
command, allowing the user to not have any extra output except the scale
factor.