Options: warning printed when option value is empty string
Until now, when the user gave a value to a long option using the '=' sign
(for example '--hdu=1'), the value should immediately come after the '='
sign. However, it can happen that the user mistakenly puts a space after
the equal sign (for example '--hdu= 1'). In such situations, the program
will usually crash with a confusing error message, or even worse, produce
unreasonable outputs. So Gnuastro's program should print a warning in such
cases to let the user know how to fix the problem.

With this commit, such a warning is now printed! When the value of an
option is an empty string, a complete error message is printed that will
let the user know about a potential problem.

This bug was reported by Faezeh Bijarchian.

This fixes bug #62944.