Fits: initialized ndim variable in new pixelareaarcsec2
Until now, the 'ndim' variable in the newly added function to calculate the
pixel area in arcsec-squared wasn't initialized (to zero). This was not a
problem because the value will be written in later. However, some compilers
printed an "used uninitialized in this function [-Wmaybe-uninitialized]"
warning.

With this commit, to avoid this wrong warning, the value has been
initalized.

This issue was reported by Faezeh Bijarchian.