Make extensions: first draft implementation
Until now, to use Gnuastro in Makefiles, it was necessary to use the
'shell' function of Make. But 'shell' requires a system call by Make, and
the called programs sometimes have a lot of overhead (when called many
times for example).

With this commit, the infrastructure has been put in place to add GNU Make
functions in Gnuastro, with one function to start with. In time, we'll be
adding more functions.