Make extensions: fixed prefix is now ast-
Until now (since last commit), the fixed prefix to Gnuastro's Make
extensions was 'astgmk'. But that consumes too many characters and make the
commands long.

With this commit, it has been changed to 'ast-', which is more similar to
the programs and familiar to the users. Some minor errors in the book (in
the menues) have also been corrected.