Make extension: new function to extract unique values of a keyword
Until now, when you wanted to get all the unique values that are given to a
FITS keyword in various files, it was only possible by calling 'astfits',
and using Table's column arithmetic to extract the unique values. Within a
shell script this is fine, but within a Makefile, this is too annoying and
can slow down the process.

With this commit, a new Make extension function of Gnuastro has been
defined for the purpose above. It can greatly help in designing generic
pipelines on data where you don't know the full set of possible keyword
values a-priori.