Table: warning if --catcolumns is given without --catcolumnfile
Until now, if '--catcolumns' (or '-C') was called, but without
'--catcolumnfile' (or '-L'), then the value given to '--catcolumns' would
be silently ignored. This could cause confusion for the user. Especially
since the use may have wanted '-c' and mistakenly written '-C'!

With this commit, in such cases, a warning is printed by Table to standard
error to let the user know about this option being ignored.

This issue was suggested by Sepideh Eskandarlou.