Book (Make extensions): better explanation in Make's variable types
Until now, the cartouch that described the difference between '=' and ':='
in Make, didn't mention that the second type can only be used after their
definition.

With this commit, this explanation has been added and generally it has been
edited to be more clear. I also noticed that the new menu entries in the
general tutorial didn't have a summary, so a summary has been added.