Library (wcs.h): conversion functions always return, NaN when bad
Until now, if only a single one of the input coordiantes was not valid (or
had any other kind of WCSLIB related error), the two 'gal_wcs_img_to_word'
and 'gal_wcs_world_to_img' functions would abort the program. However, it
often happens that the user doesn't care about the bad points and can later
remove them!

With this commit, we do a coordinate-by-coordinate check and set the bad
coordinates to NaN. This lets the users select if they want to keep or
remove that particular coordinate.