Library (python.c): Added module for helping build Python package
Until now, Gnuastro library facilitated C/C++ programs, but with this
commit we start working towards setting up the Gnuastro Python package.
The Python package will provide a high-level abstraction of the current
library functions alongwith using NumPy as the default container for any
data arrays.

With this commit, we add an extra option to the configure script to toggle
Python library module. We also add checks to conditionally build the python
library module, only if the user has Python 3 and NumPy installed.

Currently, the new module added, only provides two conversion functions
between Gnuastro and NumPy's data types. However, the aim is to collect all
utility functions that might be needed for writing the Python extensions
modules (source of the Python package) in this one place.

Also, added description of the module and its functions in the book under
the Library section.

Also, fixed a typo found in the lib/fits.c file.