Library (eps.h): Linewidth not re-written when given
Until now, when the linewidth was given, it would be mistakenly reset when
it was not different from the other marks.

With this commit, the argument that is used to write the line width is
given the proper line width and accounting of the line width in other steps
is done through the 'lwidthv' variable (that was causing a re-write).

This issue was reported by Sepideh Eskandarlou.