Library (python.h): functions renamed, configure warnings, doc edits
Until now, the names of the two Python functions
('gal_py_npytype_to_galtype' and 'gal_py_galtype_to_npytype') were a little
cryptic and repated the term 'type' two times. This makes the functions
hard to group when looking in the header, and of course in the manual, or
when using editors that automatically fill the functions names. Gnuastro's
way of grouping functions by their name is to put their common parts at the
start. In this scenario: start both with 'gal_py_type' prefix.

With this commit, this has been done and the functions are now called
'gal_py_type_to_numpy' and 'gal_py_type_from_numpy'. In the end, one of the
types is a Gnuastro type, so there is no point in repeating Gnuastro
('gal', which is also at the start of the function name).

In the process, the following changes have also been made:

 - The two functions have been added in the NEWS file.

 - The section title "Python library" has been renamed to "Python
   interface" to avoid confusion with a fully-fledged Python library: the
   goal here is just to prive an interface to communicate with
   fully-fledged Python libraries. It has also been brought up in the list
   of the Gnuastro library section to be closer to other wrappers (like
   Git).

 - The starting paragraph of the Python interface section now has a better
   introduction on what Python is, and how it relates to Gnuastro.

 - The copyright notices of the new files have been put in the standard
   format of all Gnuastro's C source files (and Jash has been set as the
   original author.

 - The '#ifdef HAVE_PYTHON' has been removed from the main body of
   functions in 'lib/python.c' because this file is only compiled if a
   working Numpy+Python3 was found.

 - The indentation of the 'case' statement in the two Python interface
   functions have beed edited to follow GNU's style.

 - All comments in the format of '//' have been removed! In Gnuastro, we
   only use the '/* */' commenting style.

 - Warnings are printed by './configure' if the user doesn't have Python
   (just like other dependencies): one short warning if a mandatory
   dependency is missing, and a longer one at the end (which is only for
   optional dependencies).

 - The last two types of 'gal_py_type_from_numpy' were mistakely taking
   Gnuastro's type identifier and returning Numpy's.