Library (python.h):  second component of function names to python
Until now, the second component of the newly added Python installed
functions in Gnuastro had the format of 'gal_py_XXXX'! However, the name of
the C source code and header are 'python.c' and 'python.h'.

With this commit, the names have been corrected to 'gal_python_XXXX' to
follow the same convention as the rest of Gnuastro's installed functions in
the library.