Library (eps.h): now possible to set the color of the border
Until now, the color of the border produced in vector graphics outputs
(like EPS or PDF) was fixed (black). But when producing figures, it is
sometimes necessary to customize the border's color.

With this commit, a new argument has been added to the 'gal_eps_write'
function to allow the caller to set the color of the border. Using this
feature, the ConvertType program now has an option called '--bordercolor'
to let the user easily select their desired color for the border. The
default value for the border color is black (in ConvertType's system-wide
configuration file).