Book (Convolve): example added on 3D convolution
Until now, the examples at the start of the "Invoking astconvolve" section
of the book didn't contain any demo for 3D convolution. Also, there was no
mention of the relevance of tessellation in convolution.

With this commit, an example 3D convolution command has been added in the
first set of examples and in the description of spatial domain convolution
a paragraph has been added to explain what role tessellation plays in it.

This was suggested by Faezeh Bijarchian.