Make (ast-text-contains): new func. to find words containing substring
Until now, it was very hard in Make to select all the words in a list that
contain a certain string! The '$(filter ...)' function is only good if your
desired string is at the start or end of the string, it can't be anywhere!
In other words, Make's '$(filter ...)' function only takes one '%'! So it
was hard to find words that may contain a certain string at any place
within the word.

With this commit, a Gnuastro Make extension function called
'$(ast-text-contains ...)' function has been added for this purpose and
simplify scenarios like above.

In the process, I also noticed that the two currently existing functions
weren't following the Make function convention of putting the list at the
end! In the two FITS-related functions, the list of files was placed at the
start, making the main parameter (the keyword/value) go to the end and
harder to see when reading the code. So with this commit, this problem has
also been fixed and they now take the keyword/value as the first
argument(s), the HDU as the second-last argument and the list of FITS files
as the last argument.